Remove the GRUB_FORCE_HIDDEN_MENU and enable the script's functionality be default.
The initial idea was to submit this upstream and see if the grub-dev
would include this as part of the base package, and that's why there
was an extra flag.

A separate branch has been kept for this potencial future, while the
default will be to enable this feature if the script is installed.